// frontend/src/app/core/api/generated/model/supplier-response.ts
/**
 * Inventory API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SupplierResponse { 
    id?: string;
    code?: string;
    legalName?: string;
    commercialName?: string;
    fiscalIdentifier?: string;
    email?: string;
    phone?: string;
    active?: boolean;
    createdAt?: string;
    updatedAt?: string;
}